+++ modules/src/main/java/com/atlassian/plugin/connect/modules/beans/WebItemTargetBean.java
@@ -101,6 +101,7 @@ public class WebItemTargetBean extends BeanWithParams
     public int hashCode()
     {
         return new HashCodeBuilder(19, 23)
+                .appendSuper(super.hashCode())
                 .append(type)
                 .append(options)
                 .build();