+++ modules/src/main/java/com/atlassian/plugin/connect/modules/beans/nested/AddOnScopeBeans.java
@@ -7,17 +7,22 @@ public class AddOnScopeBeans
     private Collection<AddOnScopeBean> scopes; // set by gson
     private Collection<AddOnScopeBean.RestPathBean> restPaths; // set by gson
     private Collection<AddOnScopeBean.SoapRpcPathBean> soapRpcPaths; // set by gson
+    private Collection<AddOnScopeBean.JsonRpcPathBean> jsonRpcPaths; // set by gson
 
     public AddOnScopeBeans()
     {
-        this(null, null, null);
+        this(null, null, null, null);
     }
 
-    public AddOnScopeBeans(Collection<AddOnScopeBean> scopes, Collection<AddOnScopeBean.RestPathBean> restPaths, Collection<AddOnScopeBean.SoapRpcPathBean> soapRpcPaths)
+    public AddOnScopeBeans(Collection<AddOnScopeBean> scopes,
+                           Collection<AddOnScopeBean.RestPathBean> restPaths,
+                           Collection<AddOnScopeBean.SoapRpcPathBean> soapRpcPaths,
+                           Collection<AddOnScopeBean.JsonRpcPathBean> jsonRpcPaths)
     {
         this.scopes = scopes;
         this.restPaths = restPaths;
         this.soapRpcPaths = soapRpcPaths;
+        this.jsonRpcPaths = jsonRpcPaths;
     }
 
     public Collection<AddOnScopeBean> getScopes()
@@ -33,5 +38,10 @@ public class AddOnScopeBeans
     public Collection<AddOnScopeBean.SoapRpcPathBean> getSoapRpcPaths()
     {
         return soapRpcPaths;
+    }
+
+    public Collection<AddOnScopeBean.JsonRpcPathBean> getJsonRpcPaths()
+    {
+        return jsonRpcPaths;
     }
 }