+++ plugin/src/test/java/com/atlassian/plugin/connect/test/plugin/capabilities/descriptor/ConnectViewProfileTabPanelModuleDescriptorFactoryTest.java
@@ -1,0 +1,13 @@
+package com.atlassian.plugin.connect.test.plugin.capabilities.descriptor;
+
+import com.atlassian.plugin.connect.plugin.capabilities.provider.ConnectTabPanelModuleProvider;
+
+
+public class ConnectViewProfileTabPanelModuleDescriptorFactoryTest extends AbstractConnectTabPanelModuleDescriptorFactoryTest
+{
+    public ConnectViewProfileTabPanelModuleDescriptorFactoryTest()
+    {
+        super(ConnectTabPanelModuleProvider.FIELD_TO_HINTS.get(ConnectTabPanelModuleProvider.PROFILE_TAB_PANELS));
+    }
+
+}