+++ docs/public/resources/release-notes.md
@@ -2,7 +2,10 @@
 # Release Notes
 
 ## [1.0-m29](../release-notes/1-0-m29.html)
-* Tabs can now be added to Confluence Space Tools section. Check out [Space Tools Tab](modules/confluence/space-tools-tab.html) documentation for more information.
+* Tabs can now be added to Confluence Space Tools section. Check out [Space Tools Tab](../modules/confluence/space-tools-tab.html)
+documentation for more information.
+* Support for [web sections](../modules/jira/web-section.html).
+* Support for full screen dialogs.
 
 Read the [1.0-m29 release notes](../release-notes/1-0-m29.html).